+++ configs/credit_assign/catch/rnn.yml
